+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `update` command for updating `--name`, `--description`, `--privacy`, and `--attribution` of a tileset
 - Add `--attribution` option to the `create` command
 - Add `delete` command for deleting a tileset
+- Add `tilejson` command
 
 # 1.0.1.dev0 (2020-04-07)
 - Fixed http status code for tilesets sources delete so it will no longer error